Completely remove old and bw/c code
