+++ nucliadb/src/nucliadb/common/datamanagers/cluster.py
@@ -17,28 +17,26 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
-import logging
-
-from nucliadb.common.datamanagers import kb_v2
 from nucliadb.common.maindb.driver import Transaction
 from nucliadb_protos import writer_pb2
 
-from .utils import datamanagers_v2_read, datamanagers_v2_write, get_kv_pb
-
-logger = logging.getLogger(__name__)
-
-
-KB_SHARDS = "/kbs/{kbid}/shards"
+from .utils import _pg_cursor
 
 
 async def get_kb_shards(
     txn: Transaction, *, kbid: str, for_update: bool = False
 ) -> writer_pb2.Shards | None:
-    if datamanagers_v2_read(kbid):
-        return await kb_v2.get_kb_shards(txn, kbid=kbid, for_update=for_update)
-
-    key = KB_SHARDS.format(kbid=kbid)
-    return await get_kv_pb(txn, key, writer_pb2.Shards, for_update=for_update)
+    async with _pg_cursor(txn) as cur:
+        statement = "SELECT shards FROM kbs WHERE kbid = %(kbid)s"
+        if for_update:
+            statement += " FOR UPDATE"
+        await cur.execute(statement, {"kbid": kbid})
+        row = await cur.fetchone()
+        if row is None or row[0] is None:
+            return None
+        pb = writer_pb2.Shards()
+        pb.ParseFromString(row[0])
+        return pb
 
 
 async def is_kb_shard(txn: Transaction, *, kbid: str, shard_id: str) -> bool:
@@ -52,8 +50,13 @@ async def is_kb_shard(txn: Transaction, *, kbid: str, shard_id: str) -> bool:
 
 
 async def update_kb_shards(txn: Transaction, *, kbid: str, shards: writer_pb2.Shards) -> None:
-    key = KB_SHARDS.format(kbid=kbid)
-    await txn.set(key, shards.SerializeToString())
-
-    if datamanagers_v2_write(kbid):
-        await kb_v2.update_kb_shards(txn, kbid=kbid, shards=shards)
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            INSERT INTO kbs (kbid, shards)
+            VALUES (%(kbid)s, %(shards)s)
+            ON CONFLICT (kbid) DO UPDATE SET
+                shards = EXCLUDED.shards
+            """,
+            {"kbid": kbid, "shards": shards.SerializeToString()},
+        )
