+++ nucliadb/src/nucliadb/common/datamanagers/conversations.py
@@ -17,58 +17,40 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
-from nucliadb.common.datamanagers import conversations_v2
-from nucliadb.common.datamanagers.utils import datamanagers_v2_read, datamanagers_v2_write
+"""
+Datamanager for the `kb_conversations` PostgreSQL table (migration 0016).
+
+Each row stores one page of a conversation field, or the SplitsMetadata sentinel:
+  - kbid     - FK → kb_resources.kbid (ON DELETE CASCADE)
+  - rid      - FK → kb_resources.rid  (ON DELETE CASCADE)
+  - field_id - user-defined conversation field name
+  - page     - 1-based page number; the sentinel value 0 stores SplitsMetadata
+  - value    - serialised protobuf bytes:
+                 page = 0  → resources_pb2.SplitsMetadata
+                 page >= 1 → resources_pb2.Conversation (~200 messages each)
+
+The FieldConversation metadata (page count, total, page size, extract/split strategy)
+is stored in kb_fields.value for the corresponding 'c'-type field row and is written
+directly here to avoid a cross-module import cycle with fields_v2.
+
+NOTE: deleting a kb_resources row (or its parent kbs row) automatically removes all
+related kb_conversations rows via the ON DELETE CASCADE foreign key.
+Deleting a single conversation field requires an explicit DELETE by (kbid, rid, field_id)
+because there is no FK from kb_conversations to kb_fields.
+"""
+
+from nucliadb.common.datamanagers.utils import _pg_cursor, logs_foreign_key_error
 from nucliadb.common.maindb.driver import Transaction
 from nucliadb_protos.resources_pb2 import Conversation as PBConversation
 from nucliadb_protos.resources_pb2 import FieldConversation, SplitsMetadata
 
-KB_CONVERSATION_PAGE = "/kbs/{kbid}/r/{uuid}/f/{type}/{field}/{page}"
-KB_CONVERSATION_SPLITS_METADATA = "/kbs/{kbid}/r/{uuid}/f/{type}/{field}/splits_metadata"
-KB_CONVERSATION_METADATA = "/kbs/{kbid}/r/{uuid}/f/{type}/{field}"
+# Sentinel page number used to store SplitsMetadata in the kb_conversations table.
+# Real conversation pages are 1-based, so 0 is always free.
+_SPLITS_METADATA_PAGE = 0
 
-
-async def get_page(
-    txn: Transaction,
-    *,
-    kbid: str,
-    rid: str,
-    field_type: str,
-    field_id: str,
-    page: int,
-) -> PBConversation | None:
-    if page <= 0:
-        raise ValueError("Conversation pages start at index 1")
-
-    if datamanagers_v2_read(kbid):
-        return await conversations_v2.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=page)
-
-    key = KB_CONVERSATION_PAGE.format(kbid=kbid, uuid=rid, type=field_type, field=field_id, page=page)
-    payload = await txn.get(key)
-    if payload is None:
-        return None
-    pb = PBConversation()
-    pb.ParseFromString(payload)
-    return pb
-
-
-async def set_page(
-    txn: Transaction,
-    *,
-    kbid: str,
-    rid: str,
-    field_type: str,
-    field_id: str,
-    page: int,
-    value: PBConversation,
-) -> None:
-    key = KB_CONVERSATION_PAGE.format(kbid=kbid, uuid=rid, type=field_type, field=field_id, page=page)
-    await txn.set(key, value.SerializeToString())
-
-    if datamanagers_v2_write(kbid):
-        await conversations_v2.set_page(
-            txn, kbid=kbid, rid=rid, field_id=field_id, page=page, value=value
-        )
+# ---------------------------------------------------------------------------
+# FieldConversation metadata  (stored in kb_fields, field_type = 'c')
+# ---------------------------------------------------------------------------
 
 
 async def get_metadata(
@@ -76,52 +58,120 @@ async def get_metadata(
     *,
     kbid: str,
     rid: str,
-    field_type: str,
     field_id: str,
 ) -> FieldConversation | None:
-    if datamanagers_v2_read(kbid):
-        return await conversations_v2.get_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
+    """Return the FieldConversation metadata for a conversation field, or None."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            SELECT value FROM kb_fields
+            WHERE kbid = %(kbid)s AND rid = %(rid)s
+              AND field_type = 'c' AND field_id = %(field_id)s
+            """,
+            {"kbid": kbid, "rid": rid, "field_id": field_id},
+        )
+        row = await cur.fetchone()
+        if row is None or row[0] is None:
+            return None
+        pb = FieldConversation()
+        pb.ParseFromString(bytes(row[0]))
+        return pb
 
-    key = KB_CONVERSATION_METADATA.format(kbid=kbid, uuid=rid, type=field_type, field=field_id)
-    payload = await txn.get(key)
-    if payload is None:
-        return None
-    pb = FieldConversation()
-    pb.ParseFromString(payload)
-    return pb
 
-
+@logs_foreign_key_error
 async def set_metadata(
     txn: Transaction,
     *,
     kbid: str,
     rid: str,
-    field_type: str,
     field_id: str,
     metadata: FieldConversation,
 ) -> None:
-    key = KB_CONVERSATION_METADATA.format(kbid=kbid, uuid=rid, type=field_type, field=field_id)
-    await txn.set(key, metadata.SerializeToString())
-
-    if datamanagers_v2_write(kbid):
-        await conversations_v2.set_metadata(
-            txn, kbid=kbid, rid=rid, field_id=field_id, metadata=metadata
+    """Upsert the FieldConversation metadata row in kb_fields."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            INSERT INTO kb_fields (kbid, rid, field_type, field_id, value)
+            VALUES (%(kbid)s, %(rid)s, 'c', %(field_id)s, %(value)s)
+            ON CONFLICT (kbid, rid, field_type, field_id) DO UPDATE SET
+                value = EXCLUDED.value
+            """,
+            {
+                "kbid": kbid,
+                "rid": rid,
+                "field_id": field_id,
+                "value": metadata.SerializeToString(),
+            },
         )
 
 
-async def delete_field(
+# ---------------------------------------------------------------------------
+# Conversation pages  (stored in kb_conversations, page >= 1)
+# ---------------------------------------------------------------------------
+
+
+async def get_page(
     txn: Transaction,
     *,
     kbid: str,
     rid: str,
-    field_type: str,
     field_id: str,
-) -> None:
-    base_key = KB_CONVERSATION_METADATA.format(kbid=kbid, uuid=rid, type=field_type, field=field_id)
-    await txn.delete_by_prefix(base_key)
+    page: int,
+) -> PBConversation | None:
+    """Return one page of a conversation field, or None if not found."""
+    if page <= 0:
+        raise ValueError("Conversation pages start at index 1")
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            SELECT value FROM kb_conversations
+            WHERE kbid = %(kbid)s AND rid = %(rid)s
+              AND field_type = 'c' AND field_id = %(field_id)s AND page = %(page)s
+            """,
+            {"kbid": kbid, "rid": rid, "field_id": field_id, "page": page},
+        )
+        row = await cur.fetchone()
+        if row is None or row[0] is None:
+            return None
+        pb = PBConversation()
+        pb.ParseFromString(bytes(row[0]))
+        return pb
 
-    if datamanagers_v2_write(kbid):
-        await conversations_v2.delete_field(txn, kbid=kbid, rid=rid, field_id=field_id)
+
+@logs_foreign_key_error
+async def set_page(
+    txn: Transaction,
+    *,
+    kbid: str,
+    rid: str,
+    field_id: str,
+    page: int,
+    value: PBConversation,
+) -> None:
+    """Upsert one page of a conversation field."""
+    if page <= 0:
+        raise ValueError("Conversation pages start at index 1")
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            INSERT INTO kb_conversations (kbid, rid, field_type, field_id, page, value)
+            VALUES (%(kbid)s, %(rid)s, 'c', %(field_id)s, %(page)s, %(value)s)
+            ON CONFLICT (kbid, rid, field_type, field_id, page) DO UPDATE SET
+                value = EXCLUDED.value
+            """,
+            {
+                "kbid": kbid,
+                "rid": rid,
+                "field_id": field_id,
+                "page": page,
+                "value": value.SerializeToString(),
+            },
+        )
+
+
+# ---------------------------------------------------------------------------
+# SplitsMetadata  (stored in kb_conversations at sentinel page = 0)
+# ---------------------------------------------------------------------------
 
 
 async def get_splits_metadata(
@@ -129,34 +179,77 @@ async def get_splits_metadata(
     *,
     kbid: str,
     rid: str,
-    field_type: str,
     field_id: str,
 ) -> SplitsMetadata | None:
-    if datamanagers_v2_read(kbid):
-        return await conversations_v2.get_splits_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
+    """Return the SplitsMetadata for a conversation field, or None."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            SELECT value FROM kb_conversations
+            WHERE kbid = %(kbid)s AND rid = %(rid)s
+              AND field_type = 'c' AND field_id = %(field_id)s AND page = %(page)s
+            """,
+            {"kbid": kbid, "rid": rid, "field_id": field_id, "page": _SPLITS_METADATA_PAGE},
+        )
+        row = await cur.fetchone()
+        if row is None or row[0] is None:
+            return None
+        pb = SplitsMetadata()
+        pb.ParseFromString(bytes(row[0]))
+        return pb
 
-    key = KB_CONVERSATION_SPLITS_METADATA.format(kbid=kbid, uuid=rid, type=field_type, field=field_id)
-    payload = await txn.get(key)
-    if payload is None:
-        return None
-    pb = SplitsMetadata()
-    pb.ParseFromString(payload)
-    return pb
 
-
+@logs_foreign_key_error
 async def set_splits_metadata(
     txn: Transaction,
     *,
     kbid: str,
     rid: str,
-    field_type: str,
     field_id: str,
     splits_metadata: SplitsMetadata,
 ) -> None:
-    key = KB_CONVERSATION_SPLITS_METADATA.format(kbid=kbid, uuid=rid, type=field_type, field=field_id)
-    await txn.set(key, splits_metadata.SerializeToString())
+    """Upsert the SplitsMetadata sentinel row (page = 0) for a conversation field."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            INSERT INTO kb_conversations (kbid, rid, field_type, field_id, page, value)
+            VALUES (%(kbid)s, %(rid)s, 'c', %(field_id)s, %(page)s, %(value)s)
+            ON CONFLICT (kbid, rid, field_type, field_id, page) DO UPDATE SET
+                value = EXCLUDED.value
+            """,
+            {
+                "kbid": kbid,
+                "rid": rid,
+                "field_id": field_id,
+                "page": _SPLITS_METADATA_PAGE,
+                "value": splits_metadata.SerializeToString(),
+            },
+        )
 
-    if datamanagers_v2_write(kbid):
-        await conversations_v2.set_splits_metadata(
-            txn, kbid=kbid, rid=rid, field_id=field_id, splits_metadata=splits_metadata
+
+# ---------------------------------------------------------------------------
+# Delete
+# ---------------------------------------------------------------------------
+
+
+async def delete_field(
+    txn: Transaction,
+    *,
+    kbid: str,
+    rid: str,
+    field_id: str,
+) -> None:
+    """
+    Delete all kb_conversations rows for this field (all pages + the
+    SplitsMetadata sentinel).  The kb_fields row for the 'c'-type field
+    is handled by the fields datamanager; cascading from kb_resources only
+    fires when the whole resource or KB is deleted.
+    """
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            DELETE FROM kb_conversations
+            WHERE kbid = %(kbid)s AND rid = %(rid)s AND field_type = 'c' AND field_id = %(field_id)s
+            """,
+            {"kbid": kbid, "rid": rid, "field_id": field_id},
         )
