+++ nucliadb/src/nucliadb/common/datamanagers/kb.py
@@ -17,107 +17,140 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
+"""
+Datamanager for the `kbs` PostgreSQL table (migration 0016).
+
+Each row represents one knowledge box and stores:
+  - kbid    - primary key
+  - slug    - human-readable unique identifier (nullable)
+  - title   - display name (nullable)
+  - shards  - serialised writer_pb2.Shards protobuf
+  - config  - serialised knowledgebox_pb2.KnowledgeBoxConfig protobuf
+"""
+
 import logging
 from collections.abc import AsyncIterator
 
-from nucliadb.common.datamanagers import kb_v2
+import psycopg.errors
+
 from nucliadb.common.datamanagers.exceptions import KnowledgeBoxNotFound
-from nucliadb.common.datamanagers.utils import datamanagers_v2_read, datamanagers_v2_write
+from nucliadb.common.datamanagers.utils import _pg_cursor
 from nucliadb.common.maindb.driver import Transaction
 from nucliadb_protos import knowledgebox_pb2
 
 from . import cluster
 
-KB_UUID = "/kbs/{kbid}/config"
-KB_SLUGS_BASE = "/kbslugs/"
-KB_SLUGS = KB_SLUGS_BASE + "{slug}"
-
 logger = logging.getLogger(__name__)
 
 
-async def get_kbs(txn: Transaction, *, prefix: str = "") -> AsyncIterator[tuple[str, str]]:
-    if datamanagers_v2_read("kbs"):
-        async for kbid, slug in kb_v2.get_kbs(txn, slug_prefix=prefix):
-            yield (kbid, slug)
-        return
-
-    async for key in txn.keys(KB_SLUGS.format(slug=prefix)):
-        slug = key.replace(KB_SLUGS_BASE, "")
-        uuid = await get_kb_uuid(txn, slug=slug)
-        if uuid is None:
-            logger.error(f"KB with slug ({slug}) but without uuid?")
-            continue
-        yield (uuid, slug)
+async def get_kbs(txn: Transaction, *, slug_prefix: str = "") -> AsyncIterator[tuple[str, str]]:
+    async with _pg_cursor(txn) as cur:
+        if slug_prefix:
+            await cur.execute(
+                "SELECT kbid, slug FROM kbs WHERE slug LIKE %(prefix)s ORDER BY kbid",
+                {"prefix": slug_prefix + "%"},
+            )
+        else:
+            await cur.execute(
+                "SELECT kbid, slug FROM kbs ORDER BY kbid",
+            )
+        async for row in cur:
+            if row[1] is not None:  # Only yield KBs that have a slug (i.e., not soft-deleted)
+                yield (str(row[0]), row[1])
 
 
 async def exists_kb(txn: Transaction, *, kbid: str) -> bool:
-    if datamanagers_v2_read("kbs"):
-        return await kb_v2.exists_kb(txn, kbid=kbid)
-
-    return await get_config(txn, kbid=kbid, for_update=False) is not None
+    async with _pg_cursor(txn) as cur:
+        try:
+            await cur.execute(
+                """
+                SELECT 1 FROM kbs
+                WHERE kbid = %(kbid)s
+                  AND slug IS NOT NULL
+                  AND deleted_at IS NULL
+                """,
+                {"kbid": kbid},
+            )
+        except psycopg.errors.InvalidTextRepresentation:
+            logger.warning(
+                "Invalid UUID format in exists_kb() check, returning False",
+                extra={"kbid": kbid},
+            )
+            return False
+        return await cur.fetchone() is not None
 
 
 async def get_kb_uuid(txn: Transaction, *, slug: str) -> str | None:
-    if datamanagers_v2_read("kbs"):
-        return await kb_v2.get_kb_uuid(txn, slug=slug)
-
-    uuid = await txn.get(KB_SLUGS.format(slug=slug), for_update=False)
-    if uuid is not None:
-        return uuid.decode()
-    else:
-        return None
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "SELECT kbid FROM kbs WHERE slug = %(slug)s",
+            {"slug": slug},
+        )
+        row = await cur.fetchone()
+        return str(row[0]) if row is not None else None
 
 
 async def set_kbid_for_slug(txn: Transaction, *, slug: str, kbid: str):
-    key = KB_SLUGS.format(slug=slug)
-    await txn.set(key, kbid.encode())
-
-    if datamanagers_v2_write("kbs"):
-        await kb_v2.set_kbid_for_slug(txn, slug=slug, kbid=kbid)
-
-
-async def modify_slug(txn: Transaction, *, kbid: str, old_slug: str, new_slug: str) -> None:
-    await txn.delete(KB_SLUGS.format(slug=old_slug))
-    await txn.set(
-        KB_SLUGS.format(slug=new_slug),
-        kbid.encode(),
-    )
-
-    if datamanagers_v2_write("kbs") or datamanagers_v2_write(kbid):
-        await kb_v2.set_kbid_for_slug(txn, slug=new_slug, kbid=kbid)
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            INSERT INTO kbs (kbid, slug)
+            VALUES (%(kbid)s, %(slug)s)
+            ON CONFLICT (kbid) DO UPDATE SET
+                slug = EXCLUDED.slug
+            """,
+            {"kbid": kbid, "slug": slug},
+        )
 
 
-async def delete_kb_slug(txn: Transaction, *, slug: str):
-    key = KB_SLUGS.format(slug=slug)
-    await txn.delete(key)
+async def delete(txn: Transaction, *, kbid: str) -> None:
+    """Fully delete a KB row and all its associated resources, fields, and conversations."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "DELETE FROM kbs WHERE kbid = %(kbid)s",
+            {"kbid": kbid},
+        )
+
+
+async def soft_delete(txn: Transaction, *, kbid: str) -> None:
+    """Soft delete a KB row by clearing its slug and stamping deleted_at with the current time.
+
+    No-op if the KB does not exist (UPDATE affects 0 rows without raising an error).
+    """
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "UPDATE kbs SET slug = NULL, deleted_at = NOW() WHERE kbid = %(kbid)s",
+            {"kbid": kbid},
+        )
 
 
 async def get_config(
     txn: Transaction, *, kbid: str, for_update: bool = False
 ) -> knowledgebox_pb2.KnowledgeBoxConfig | None:
-    if datamanagers_v2_read(kbid) or datamanagers_v2_read("kbs"):
-        return await kb_v2.get_config(txn, kbid=kbid)
-
-    key = KB_UUID.format(kbid=kbid)
-    payload = await txn.get(key, for_update=for_update)
-    if payload is None:
-        return None
-    response = knowledgebox_pb2.KnowledgeBoxConfig()
-    response.ParseFromString(payload)
-    return response
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "SELECT config FROM kbs WHERE kbid = %(kbid)s",
+            {"kbid": kbid},
+        )
+        row = await cur.fetchone()
+        if row is None or row[0] is None:
+            return None
+        pb = knowledgebox_pb2.KnowledgeBoxConfig()
+        pb.ParseFromString(row[0])
+        return pb
 
 
 async def set_config(txn: Transaction, *, kbid: str, config: knowledgebox_pb2.KnowledgeBoxConfig):
-    key = KB_UUID.format(kbid=kbid)
-    await txn.set(key, config.SerializeToString())
-
-    if datamanagers_v2_write(kbid) or datamanagers_v2_write("kbs"):
-        await kb_v2.set_config(txn, kbid=kbid, config=config)
-
-
-async def delete_config(txn: Transaction, *, kbid: str) -> None:
-    key = KB_UUID.format(kbid=kbid)
-    await txn.delete(key)
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            INSERT INTO kbs (kbid, config)
+            VALUES (%(kbid)s, %(config)s)
+            ON CONFLICT (kbid) DO UPDATE SET
+                config = EXCLUDED.config
+            """,
+            {"kbid": kbid, "config": config.SerializeToString()},
+        )
 
 
 async def get_model_metadata(txn: Transaction, *, kbid: str) -> knowledgebox_pb2.SemanticModelMetadata:
