+++ nucliadb/src/nucliadb/common/datamanagers/resources.py
@@ -17,343 +17,373 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
-from collections.abc import AsyncGenerator
+"""
+Datamanager for the `kb_resources` PostgreSQL table (migration 0016).
 
-import backoff
+Each row represents one resource in a knowledge box and stores:
+  - kbid           - FK → kbs.kbid (ON DELETE RESTRICT)
+  - rid            - resource UUID
+  - slug           - optional human-readable identifier
+  - shard          - shard ID the resource belongs to
+  - basic          - serialised resources_pb2.Basic
+  - origin         - serialised resources_pb2.Origin
+  - security       - serialised resources_pb2.Security
+  - extra          - serialised resources_pb2.Extra
+"""
 
-from nucliadb.common.datamanagers import fields_v2, resources_v2
-from nucliadb.common.datamanagers.utils import datamanagers_v2_read, datamanagers_v2_write, get_kv_pb
+import logging
+import uuid
+from collections.abc import AsyncIterator
+
+import psycopg.errors
+
+from nucliadb.common.datamanagers.utils import _pg_cursor, logs_foreign_key_error, with_ro_transaction
 from nucliadb.common.maindb.driver import Transaction
 from nucliadb.common.maindb.exceptions import ConflictError, NotFoundError
-
-# These should be refactored
-from nucliadb.ingest.settings import settings as ingest_settings
 from nucliadb_protos import resources_pb2
 
-from .utils import with_ro_transaction
-
-KB_RESOURCE_BASIC = "/kbs/{kbid}/r/{uuid}"
-KB_RESOURCE_BASIC_FS = "/kbs/{kbid}/r/{uuid}/basic"  # Only used on FS driver
-KB_RESOURCE_ORIGIN = "/kbs/{kbid}/r/{uuid}/origin"
-KB_RESOURCE_EXTRA = "/kbs/{kbid}/r/{uuid}/extra"
-KB_RESOURCE_SECURITY = "/kbs/{kbid}/r/{uuid}/security"
-
-KB_RESOURCE_SLUG_BASE = "/kbs/{kbid}/s/"
-KB_RESOURCE_SLUG = f"{KB_RESOURCE_SLUG_BASE}{{slug}}"
-
-KB_RESOURCE_FIELDS = "/kbs/{kbid}/r/{uuid}/f/"
-
-KB_RESOURCE_ALL_FIELDS = "/kbs/{kbid}/r/{uuid}/allfields"
-KB_MATERIALIZED_RESOURCES_COUNT = "/kbs/{kbid}/materialized/resources/count"
-
-KB_RESOURCE_SHARD = "/kbs/{kbid}/r/{uuid}/shard"
+logger = logging.getLogger(__name__)
 
 
-async def resource_exists(txn: Transaction, *, kbid: str, rid: str) -> bool:
-    if datamanagers_v2_read(kbid):
-        return await resources_v2.exists(txn, kbid=kbid, rid=rid)
-
-    basic = await get_basic_raw(txn, kbid=kbid, rid=rid)
-    return basic is not None
+def _to_rid(value: uuid.UUID) -> str:
+    """Return the 32-char hex form (no hyphens) of a UUID column value."""
+    return value.hex
 
 
-# id and slug
+# ---------------------------------------------------------------------------
+# Write operations
+# ---------------------------------------------------------------------------
 
 
-async def get_resource_uuid_from_slug(txn: Transaction, *, kbid: str, slug: str) -> str | None:
-    if datamanagers_v2_read(kbid):
-        return await resources_v2.get_resource_uuid_from_slug(txn, kbid=kbid, slug=slug)
-
-    encoded_uuid = await txn.get(KB_RESOURCE_SLUG.format(kbid=kbid, slug=slug, for_update=False))
-    if not encoded_uuid:
-        return None
-    return encoded_uuid.decode()
-
-
-async def slug_exists(txn: Transaction, *, kbid: str, slug: str) -> bool:
-    if datamanagers_v2_read(kbid):
-        return await resources_v2.slug_exists(txn, kbid=kbid, slug=slug)
-
-    key = KB_RESOURCE_SLUG.format(kbid=kbid, slug=slug)
-    encoded_slug: bytes | None = await txn.get(key)
-    return encoded_slug not in (None, b"")
-
-
-async def set_slug(txn: Transaction, *, kbid: str, rid: str, slug: str) -> None:
-    key = KB_RESOURCE_SLUG.format(kbid=kbid, slug=slug)
-    await txn.set(key, rid.encode())
-
-    if datamanagers_v2_write(kbid):
-        await resources_v2.set_slug(txn, kbid=kbid, rid=rid, slug=slug)
+@logs_foreign_key_error
+async def set_basic(
+    txn: Transaction,
+    *,
+    kbid: str,
+    rid: str,
+    basic: resources_pb2.Basic,
+) -> None:
+    """Upsert the basic column of a resource row, creating the row if it does not exist."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            INSERT INTO kb_resources (kbid, rid, basic)
+            VALUES (%(kbid)s, %(rid)s, %(basic)s)
+            ON CONFLICT (kbid, rid) DO UPDATE SET
+                basic = EXCLUDED.basic
+            """,
+            {"kbid": kbid, "rid": rid, "basic": basic.SerializeToString()},
+        )
 
 
-async def modify_slug(txn: Transaction, *, kbid: str, rid: str, new_slug: str) -> str:
-    basic = await get_basic(txn, kbid=kbid, rid=rid)
-    if basic is None:
+@logs_foreign_key_error
+async def set_origin(
+    txn: Transaction,
+    *,
+    kbid: str,
+    rid: str,
+    origin: resources_pb2.Origin,
+) -> None:
+    """Upsert the origin column, creating the row if it does not exist."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            INSERT INTO kb_resources (kbid, rid, origin)
+            VALUES (%(kbid)s, %(rid)s, %(origin)s)
+            ON CONFLICT (kbid, rid) DO UPDATE SET
+                origin = EXCLUDED.origin
+            """,
+            {"kbid": kbid, "rid": rid, "origin": origin.SerializeToString()},
+        )
+
+
+@logs_foreign_key_error
+async def set_security(
+    txn: Transaction,
+    *,
+    kbid: str,
+    rid: str,
+    security: resources_pb2.Security,
+) -> None:
+    """Upsert the security column, creating the row if it does not exist."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            INSERT INTO kb_resources (kbid, rid, security)
+            VALUES (%(kbid)s, %(rid)s, %(security)s)
+            ON CONFLICT (kbid, rid) DO UPDATE SET
+                security = EXCLUDED.security
+            """,
+            {"kbid": kbid, "rid": rid, "security": security.SerializeToString()},
+        )
+
+
+@logs_foreign_key_error
+async def set_extra(
+    txn: Transaction,
+    *,
+    kbid: str,
+    rid: str,
+    extra: resources_pb2.Extra,
+) -> None:
+    """Upsert the extra column, creating the row if it does not exist."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            INSERT INTO kb_resources (kbid, rid, extra)
+            VALUES (%(kbid)s, %(rid)s, %(extra)s)
+            ON CONFLICT (kbid, rid) DO UPDATE SET
+                extra = EXCLUDED.extra
+            """,
+            {"kbid": kbid, "rid": rid, "extra": extra.SerializeToString()},
+        )
+
+
+async def get_slug(txn: Transaction, kbid: str, rid: str) -> str | None:
+    """Get the slug of a resource."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            SELECT slug FROM kb_resources
+            WHERE kbid = %(kbid)s AND rid = %(rid)s
+            """,
+            {"kbid": kbid, "rid": rid},
+        )
+        row = await cur.fetchone()
+        return str(row[0]) if row is not None else None
+
+
+@logs_foreign_key_error
+async def set_slug(
+    txn: Transaction,
+    *,
+    kbid: str,
+    rid: str,
+    slug: str,
+) -> None:
+    """Update only the slug column of an existing resource row.
+
+    Raises ConflictError if the slug already belongs to another resource in
+    the same knowledge box.
+    """
+    async with _pg_cursor(txn) as cur:
+        try:
+            await cur.execute(
+                """
+                INSERT INTO kb_resources (kbid, rid, slug)
+                VALUES (%(kbid)s, %(rid)s, %(slug)s)
+                ON CONFLICT (kbid, rid) DO UPDATE SET
+                    slug = EXCLUDED.slug
+                """,
+                {"kbid": kbid, "rid": rid, "slug": slug},
+            )
+        except psycopg.errors.UniqueViolation:
+            raise ConflictError(f"Slug '{slug}' already exists")
+
+
+@logs_foreign_key_error
+async def modify_slug(
+    txn: Transaction,
+    *,
+    kbid: str,
+    rid: str,
+    new_slug: str,
+) -> str:
+    """Update only the slug column of an existing resource row.
+
+    Returns the old slug value.
+    Raises NotFoundError if the resource does not exist.
+    Raises ConflictError if the slug already belongs to another resource in
+    the same knowledge box.
+    """
+    old_slug = await get_slug(txn, kbid=kbid, rid=rid)
+    if old_slug is None:
         raise NotFoundError()
-    old_slug = basic.slug
-
-    uuid_for_new_slug = await get_resource_uuid_from_slug(txn, kbid=kbid, slug=new_slug)
-    if uuid_for_new_slug is not None:
-        if uuid_for_new_slug == rid:
-            # Nothing to change
+    async with _pg_cursor(txn) as cur:
+        try:
+            await cur.execute(
+                """
+                UPDATE kb_resources SET slug = %(slug)s
+                WHERE kbid = %(kbid)s AND rid = %(rid)s
+                """,
+                {"kbid": kbid, "rid": rid, "slug": new_slug},
+            )
             return old_slug
-        else:
-            raise ConflictError(f"Slug {new_slug} already exists")
-
-    key = KB_RESOURCE_SLUG.format(kbid=kbid, slug=old_slug)
-    await txn.delete(key)
-    key = KB_RESOURCE_SLUG.format(kbid=kbid, slug=new_slug)
-    await txn.set(key, rid.encode())
-    basic.slug = new_slug
-    await set_basic(txn, kbid=kbid, rid=rid, basic=basic)
-
-    if datamanagers_v2_write(kbid):
-        await resources_v2.modify_slug(txn, kbid=kbid, rid=rid, new_slug=new_slug)
-
-    return old_slug
+        except psycopg.errors.UniqueViolation:
+            raise ConflictError(f"Slug '{new_slug}' already exists")
 
 
-# resource-shard
-
-
-@backoff.on_exception(backoff.expo, (Exception,), jitter=backoff.random_jitter, max_tries=3)
-async def get_resource_shard_id(
-    txn: Transaction, *, kbid: str, rid: str, for_update: bool = False
-) -> str | None:
-
-    if datamanagers_v2_read(kbid):
-        return await resources_v2.get_resource_shard_id(txn, kbid=kbid, rid=rid, for_update=for_update)
-
-    key = KB_RESOURCE_SHARD.format(kbid=kbid, uuid=rid)
-    shard = await txn.get(key, for_update=for_update)
-    if shard is not None:
-        return shard.decode()
-    else:
-        return None
-
-
-async def set_resource_shard_id(txn: Transaction, *, kbid: str, rid: str, shard: str):
-    await txn.set(KB_RESOURCE_SHARD.format(kbid=kbid, uuid=rid), shard.encode())
-
-    if datamanagers_v2_write(kbid):
-        await resources_v2.set_resource_shard_id(txn, kbid=kbid, rid=rid, shard=shard)
-
-
-# Basic
-
-
-async def get_basic(txn: Transaction, *, kbid: str, rid: str) -> resources_pb2.Basic | None:
-    if datamanagers_v2_read(kbid):
-        return await resources_v2.get_basic(txn, kbid=kbid, rid=rid)
-
-    raw = await get_basic_raw(txn, kbid=kbid, rid=rid)
-    if raw is None:
-        return None
-    basic = resources_pb2.Basic()
-    basic.ParseFromString(raw)
-    return basic
-
-
-async def get_basic_raw(txn: Transaction, *, kbid: str, rid: str) -> bytes | None:
-    if ingest_settings.driver == "local":
-        raw_basic = await txn.get(KB_RESOURCE_BASIC_FS.format(kbid=kbid, uuid=rid))
-    else:
-        raw_basic = await txn.get(KB_RESOURCE_BASIC.format(kbid=kbid, uuid=rid))
-    return raw_basic
-
-
-async def set_basic(txn: Transaction, *, kbid: str, rid: str, basic: resources_pb2.Basic):
-    if ingest_settings.driver == "local":
-        await txn.set(
-            KB_RESOURCE_BASIC_FS.format(kbid=kbid, uuid=rid),
-            basic.SerializeToString(),
+@logs_foreign_key_error
+async def set_resource_shard_id(
+    txn: Transaction,
+    *,
+    kbid: str,
+    rid: str,
+    shard: str,
+) -> None:
+    """Upsert the shard column, creating the row if it does not exist."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            """
+            INSERT INTO kb_resources (kbid, rid, shard)
+            VALUES (%(kbid)s, %(rid)s, %(shard)s)
+            ON CONFLICT (kbid, rid) DO UPDATE SET
+                shard = EXCLUDED.shard
+            """,
+            {"kbid": kbid, "rid": rid, "shard": shard},
         )
-    else:
-        await txn.set(
-            KB_RESOURCE_BASIC.format(kbid=kbid, uuid=rid),
-            basic.SerializeToString(),
-        )
-    if datamanagers_v2_write(kbid):
-        await resources_v2.set_basic(txn, kbid=kbid, rid=rid, basic=basic)
-
-
-# Origin
-
-
-async def get_origin(txn: Transaction, *, kbid: str, rid: str) -> resources_pb2.Origin | None:
-    if datamanagers_v2_read(kbid):
-        return await resources_v2.get_origin(txn, kbid=kbid, rid=rid)
-
-    key = KB_RESOURCE_ORIGIN.format(kbid=kbid, uuid=rid)
-    return await get_kv_pb(txn, key, resources_pb2.Origin, for_update=False)
-
-
-async def set_origin(txn: Transaction, *, kbid: str, rid: str, origin: resources_pb2.Origin):
-    key = KB_RESOURCE_ORIGIN.format(kbid=kbid, uuid=rid)
-    await txn.set(key, origin.SerializeToString())
-
-    if datamanagers_v2_write(kbid):
-        await resources_v2.set_origin(txn, kbid=kbid, rid=rid, origin=origin)
-
-
-# Extra
-
-
-async def get_extra(txn: Transaction, *, kbid: str, rid: str) -> resources_pb2.Extra | None:
-    if datamanagers_v2_read(kbid):
-        return await resources_v2.get_extra(txn, kbid=kbid, rid=rid)
-
-    key = KB_RESOURCE_EXTRA.format(kbid=kbid, uuid=rid)
-    return await get_kv_pb(txn, key, resources_pb2.Extra, for_update=False)
-
-
-async def set_extra(txn: Transaction, *, kbid: str, rid: str, extra: resources_pb2.Extra):
-    key = KB_RESOURCE_EXTRA.format(kbid=kbid, uuid=rid)
-    await txn.set(key, extra.SerializeToString())
-
-    if datamanagers_v2_write(kbid):
-        await resources_v2.set_extra(txn, kbid=kbid, rid=rid, extra=extra)
-
-
-# Security
-
-
-async def get_security(txn: Transaction, *, kbid: str, rid: str) -> resources_pb2.Security | None:
-    if datamanagers_v2_read(kbid):
-        return await resources_v2.get_security(txn, kbid=kbid, rid=rid)
-
-    key = KB_RESOURCE_SECURITY.format(kbid=kbid, uuid=rid)
-    return await get_kv_pb(txn, key, resources_pb2.Security, for_update=False)
-
-
-async def set_security(txn: Transaction, *, kbid: str, rid: str, security: resources_pb2.Security):
-    key = KB_RESOURCE_SECURITY.format(kbid=kbid, uuid=rid)
-    await txn.set(key, security.SerializeToString())
-
-    if datamanagers_v2_write(kbid):
-        await resources_v2.set_security(txn, kbid=kbid, rid=rid, security=security)
-
-
-# KB resource ids (this functions use internal transactions, breaking the
-# datamanager contract. We should rethink them at some point)
-
-
-async def iterate_resource_ids(*, kbid: str) -> AsyncGenerator[str, None]:
-    """
-    Currently, the implementation of this is optimizing for reducing
-    how long a transaction will be open since the caller controls
-    how long each item that is yielded will be processed.
-
-    For this reason, it is not using the `txn` argument passed in.
-    """
-    if datamanagers_v2_read(kbid):
-        async for rid in resources_v2.iterate_resource_ids(kbid=kbid):
-            yield rid
-        return
-
-    batch = []
-    async for slug in _iter_resource_slugs(kbid=kbid):
-        batch.append(slug)
-        if len(batch) >= 200:
-            for rid in await _get_resource_ids_from_slugs(kbid=kbid, slugs=batch):
-                yield rid
-            batch = []
-    if len(batch) > 0:
-        for rid in await _get_resource_ids_from_slugs(kbid=kbid, slugs=batch):
-            yield rid
-
-
-@backoff.on_exception(backoff.expo, (Exception,), jitter=backoff.random_jitter, max_tries=3)
-async def _iter_resource_slugs(*, kbid: str) -> AsyncGenerator[str, None]:
-    async with with_ro_transaction() as txn:
-        async for key in txn.keys(match=KB_RESOURCE_SLUG_BASE.format(kbid=kbid)):
-            yield key.split("/")[-1]
-
-
-@backoff.on_exception(backoff.expo, (Exception,), jitter=backoff.random_jitter, max_tries=3)
-async def _get_resource_ids_from_slugs(kbid: str, slugs: list[str]) -> list[str]:
-    async with with_ro_transaction() as txn:
-        rids = await txn.batch_get([KB_RESOURCE_SLUG.format(kbid=kbid, slug=slug) for slug in slugs])
-    return [rid.decode() for rid in rids if rid is not None]
-
-
-# KB resource count (materialized key)
-
-
-async def calculate_number_of_resources(txn: Transaction, *, kbid: str) -> int:
-    """
-    Calculate the number of resources in a knowledgebox.
-
-    This is usually not very fast at all.
-
-    Long term, we could think about implementing a counter; however,
-    right now, a counter would be difficult, require a lot of
-    refactoring and not worth much value for the APIs we need
-    this feature for.
-
-    Finally, we could also query this data from the node; however,
-    it is not the source of truth for the value so it is not ideal
-    to move it to the node.
-    """
-    if datamanagers_v2_read(kbid):
-        return await resources_v2.calculate_number_of_resources(txn, kbid=kbid)
-
-    return await txn.count(KB_RESOURCE_SLUG_BASE.format(kbid=kbid))
-
-
-async def get_number_of_resources(txn: Transaction, *, kbid: str) -> int:
-    """
-    Return cached number of resources in a knowledgebox.
-    """
-    raw_value = await txn.get(KB_MATERIALIZED_RESOURCES_COUNT.format(kbid=kbid), for_update=False)
-    if raw_value is None:
-        return -1
-    return int(raw_value)
-
-
-async def set_number_of_resources(txn: Transaction, kbid: str, value: int) -> None:
-    await txn.set(KB_MATERIALIZED_RESOURCES_COUNT.format(kbid=kbid), str(value).encode())
-
-
-# Fields (materialized key with all field ids)
-
-
-async def get_all_field_ids(
-    txn: Transaction, *, kbid: str, rid: str, for_update: bool = False
-) -> resources_pb2.AllFieldIDs | None:
-    if datamanagers_v2_read(kbid):
-        return await fields_v2.get_all_field_ids(txn, kbid=kbid, rid=rid)
-
-    key = KB_RESOURCE_ALL_FIELDS.format(kbid=kbid, uuid=rid)
-    return await get_kv_pb(txn, key, resources_pb2.AllFieldIDs, for_update=for_update)
-
-
-async def set_all_field_ids(
-    txn: Transaction, *, kbid: str, rid: str, allfields: resources_pb2.AllFieldIDs
-):
-    key = KB_RESOURCE_ALL_FIELDS.format(kbid=kbid, uuid=rid)
-    await txn.set(key, allfields.SerializeToString())
-
-
-async def has_field(txn: Transaction, *, kbid: str, rid: str, field_id: resources_pb2.FieldID) -> bool:
-    if datamanagers_v2_read(kbid):
-        return await fields_v2.has_field(txn, kbid=kbid, rid=rid, field_id=field_id)
-
-    fields = await get_all_field_ids(txn, kbid=kbid, rid=rid)
-    if fields is None:
-        return False
-    for resource_field_id in fields.fields:
-        if field_id == resource_field_id:
-            return True
-    return False
 
 
 async def delete(txn: Transaction, *, kbid: str, rid: str) -> None:
-    basic = await get_basic(txn, kbid=kbid, rid=rid)
-    if basic and basic.slug:
-        await txn.delete(KB_RESOURCE_SLUG.format(kbid=kbid, slug=basic.slug))
-    await txn.delete_by_prefix(KB_RESOURCE_BASIC.format(kbid=kbid, uuid=rid))
+    """Delete a resource row (cascades to fields)."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "DELETE FROM kb_resources WHERE kbid = %(kbid)s AND rid = %(rid)s",
+            {"kbid": kbid, "rid": rid},
+        )
 
-    if datamanagers_v2_write(kbid):
-        await resources_v2.delete(txn, kbid=kbid, rid=rid)
+
+# ---------------------------------------------------------------------------
+# Read operations
+# ---------------------------------------------------------------------------
+
+
+async def exists(txn: Transaction, *, kbid: str, rid: str) -> bool:
+    """Return True if the resource exists."""
+    async with _pg_cursor(txn) as cur:
+        try:
+            await cur.execute(
+                "SELECT 1 FROM kb_resources WHERE kbid = %(kbid)s AND rid = %(rid)s",
+                {"kbid": kbid, "rid": rid},
+            )
+        except psycopg.errors.InvalidTextRepresentation:
+            logger.warning(
+                "Invalid UUID format in exists() check, returning False",
+                extra={"kbid": kbid, "rid": rid},
+            )
+            return False
+        return await cur.fetchone() is not None
+
+
+resource_exists = exists  # alias for backwards compatibility
+
+
+async def get_resource_uuid_from_slug(txn: Transaction, *, kbid: str, slug: str) -> str | None:
+    """Return the resource UUID for the given slug within a KB, or None."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "SELECT rid FROM kb_resources WHERE kbid = %(kbid)s AND slug = %(slug)s",
+            {"kbid": kbid, "slug": slug},
+        )
+        row = await cur.fetchone()
+        return _to_rid(row[0]) if row is not None else None
+
+
+async def slug_exists(txn: Transaction, *, kbid: str, slug: str) -> bool:
+    """Return True if a resource with the given slug exists within a KB."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "SELECT 1 FROM kb_resources WHERE kbid = %(kbid)s AND slug = %(slug)s",
+            {"kbid": kbid, "slug": slug},
+        )
+        return await cur.fetchone() is not None
+
+
+async def get_basic(txn: Transaction, *, kbid: str, rid: str) -> resources_pb2.Basic | None:
+    """Return the deserialised Basic for a resource, or None."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "SELECT basic FROM kb_resources WHERE kbid = %(kbid)s AND rid = %(rid)s",
+            {"kbid": kbid, "rid": rid},
+        )
+        row = await cur.fetchone()
+        if row is None or row[0] is None:
+            return None
+        pb = resources_pb2.Basic()
+        pb.ParseFromString(bytes(row[0]))
+        return pb
+
+
+async def get_origin(txn: Transaction, *, kbid: str, rid: str) -> resources_pb2.Origin | None:
+    """Return the deserialised Origin for a resource, or None."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "SELECT origin FROM kb_resources WHERE kbid = %(kbid)s AND rid = %(rid)s",
+            {"kbid": kbid, "rid": rid},
+        )
+        row = await cur.fetchone()
+        if row is None or row[0] is None:
+            return None
+        pb = resources_pb2.Origin()
+        pb.ParseFromString(bytes(row[0]))
+        return pb
+
+
+async def get_security(txn: Transaction, *, kbid: str, rid: str) -> resources_pb2.Security | None:
+    """Return the deserialised Security for a resource, or None."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "SELECT security FROM kb_resources WHERE kbid = %(kbid)s AND rid = %(rid)s",
+            {"kbid": kbid, "rid": rid},
+        )
+        row = await cur.fetchone()
+        if row is None or row[0] is None:
+            return None
+        pb = resources_pb2.Security()
+        pb.ParseFromString(bytes(row[0]))
+        return pb
+
+
+async def get_extra(txn: Transaction, *, kbid: str, rid: str) -> resources_pb2.Extra | None:
+    """Return the deserialised Extra for a resource, or None."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "SELECT extra FROM kb_resources WHERE kbid = %(kbid)s AND rid = %(rid)s",
+            {"kbid": kbid, "rid": rid},
+        )
+        row = await cur.fetchone()
+        if row is None or row[0] is None:
+            return None
+        pb = resources_pb2.Extra()
+        pb.ParseFromString(bytes(row[0]))
+        return pb
+
+
+async def iterate_resource_ids(*, kbid: str) -> AsyncIterator[str]:
+    """Iterate over all resource UUIDs in a knowledge box."""
+    async with with_ro_transaction() as txn:
+        async with _pg_cursor(txn) as cur:
+            await cur.execute(
+                "SELECT rid FROM kb_resources WHERE kbid = %(kbid)s ORDER BY rid",
+                {"kbid": kbid},
+            )
+            async for (rid,) in cur:
+                yield _to_rid(rid)
+
+
+async def calculate_number_of_resources(txn: Transaction, *, kbid: str) -> int:
+    """Return the total number of resources in a knowledge box."""
+    async with _pg_cursor(txn) as cur:
+        await cur.execute(
+            "SELECT COUNT(*) FROM kb_resources WHERE kbid = %(kbid)s",
+            {"kbid": kbid},
+        )
+        row = await cur.fetchone()
+        return row[0] if row else 0
+
+
+async def get_number_of_resources(txn: Transaction, *, kbid: str) -> int:
+    return await calculate_number_of_resources(txn, kbid=kbid)
+
+
+async def get_resource_shard_id(
+    txn: Transaction, *, kbid: str, rid: str, for_update: bool = False
+) -> str | None:
+    """Return the shard ID for a resource, or None."""
+    async with _pg_cursor(txn) as cur:
+        sql = "SELECT shard FROM kb_resources WHERE kbid = %(kbid)s AND rid = %(rid)s"
+        if for_update:
+            sql += " FOR UPDATE"
+        await cur.execute(sql, {"kbid": kbid, "rid": rid})
+        row = await cur.fetchone()
+        return row[0] if row is not None else None
