+++ nucliadb/src/nucliadb/ingest/consumer/materializer.py
@@ -96,10 +96,6 @@ class MaterializerHandler:
         logger.info(f"Materializing knowledgebox", extra={"kbid": kbid})
         async with datamanagers.with_ro_transaction() as txn:
             value = await datamanagers.resources.calculate_number_of_resources(txn, kbid=kbid)
-        async with datamanagers.with_transaction() as txn:
-            await datamanagers.resources.set_number_of_resources(txn, kbid=kbid, value=value)
-            await txn.commit()
-
         audit = get_audit()
         if audit:
             audit.report_resources(kbid=kbid, resources=value)
