+++ nucliadb/tests/ingest/unit/fields/test_conversation.py
@@ -42,7 +42,7 @@ async def txn(maindb_driver: Driver):
 async def kbid(maindb_driver: Driver):
     kbid = KnowledgeBox.new_unique_kbid()
     async with maindb_driver.rw_transaction() as txn:
-        await datamanagers.kb.kb_v2.set_kbid_for_slug(txn, kbid=kbid, slug=f"slug-{kbid}")
+        await datamanagers.kb.set_kbid_for_slug(txn, kbid=kbid, slug=f"slug-{kbid}")
         await txn.commit()
     return kbid
 
@@ -58,7 +58,7 @@ def storage():
 async def resource(maindb_driver: Driver, txn: Transaction, storage, kbid: str):
     rid = Resource.new_unique_rid()
     async with maindb_driver.rw_transaction() as txn:
-        await datamanagers.resources.resources_v2.set_slug(txn, kbid=kbid, rid=rid, slug=f"slug-{rid}")
+        await datamanagers.resources.set_slug(txn, kbid=kbid, rid=rid, slug=f"slug-{rid}")
         await txn.commit()
 
     resource = mock.AsyncMock()
