+++ nucliadb/tests/nucliadb/integration/common/datamanagers/test_conversations_v2.py
@@ -18,7 +18,7 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 """
-Integration tests for nucliadb.common.datamanagers.conversations_v2.
+Integration tests for nucliadb.common.datamanagers.conversations.
 
 Covers: set_metadata, get_metadata, set_page, get_page,
         set_splits_metadata, get_splits_metadata, delete_field.
@@ -28,7 +28,7 @@ import logging
 
 import pytest
 
-from nucliadb.common.datamanagers import conversations_v2, fields_v2, kb_v2, resources_v2
+from nucliadb.common.datamanagers import conversations, fields, kb, resources
 from nucliadb.common.maindb.driver import Driver
 from nucliadb.ingest.orm.knowledgebox import KnowledgeBox
 from nucliadb.ingest.orm.resource import Resource
@@ -49,7 +49,7 @@ from nucliadb_protos.resources_pb2 import (
 async def kbid(maindb_driver: Driver) -> str:
     kbid = KnowledgeBox.new_unique_kbid()
     async with maindb_driver.rw_transaction() as txn:
-        await kb_v2.set_kbid_for_slug(txn, kbid=kbid, slug=f"slug-{kbid}")
+        await kb.set_kbid_for_slug(txn, kbid=kbid, slug=f"slug-{kbid}")
         await txn.commit()
     return kbid
 
@@ -58,7 +58,7 @@ async def kbid(maindb_driver: Driver) -> str:
 async def rid(maindb_driver: Driver, kbid: str) -> str:
     rid = Resource.new_unique_rid()
     async with maindb_driver.rw_transaction() as txn:
-        await resources_v2.set_slug(txn, kbid=kbid, rid=rid, slug=f"slug-{rid}")
+        await resources.set_slug(txn, kbid=kbid, rid=rid, slug=f"slug-{rid}")
         await txn.commit()
     return rid
 
@@ -68,7 +68,7 @@ async def field_id(maindb_driver: Driver, kbid: str, rid: str) -> str:
     """Create the parent kb_fields row ('c' type) so FK constraints are satisfied."""
     fid = "chat"
     async with maindb_driver.rw_transaction() as txn:
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type="c", field_id=fid, value=b"")
+        await fields.set(txn, kbid=kbid, rid=rid, field_type="c", field_id=fid, value=b"")
         await txn.commit()
     return fid
 
@@ -111,11 +111,11 @@ async def test_set_and_get_metadata(maindb_driver: Driver, kbid: str, rid: str, 
     meta = make_metadata(pages=3)
 
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.set_metadata(txn, kbid=kbid, rid=rid, field_id=field_id, metadata=meta)
+        await conversations.set_metadata(txn, kbid=kbid, rid=rid, field_id=field_id, metadata=meta)
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        result = await conversations_v2.get_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
+        result = await conversations.get_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
 
     assert result is not None
     assert result.pages == 3
@@ -126,19 +126,19 @@ async def test_set_metadata_overwrites(
     maindb_driver: Driver, kbid: str, rid: str, field_id: str
 ) -> None:
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.set_metadata(
+        await conversations.set_metadata(
             txn, kbid=kbid, rid=rid, field_id=field_id, metadata=make_metadata(pages=1)
         )
         await txn.commit()
 
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.set_metadata(
+        await conversations.set_metadata(
             txn, kbid=kbid, rid=rid, field_id=field_id, metadata=make_metadata(pages=5)
         )
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        result = await conversations_v2.get_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
+        result = await conversations.get_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
 
     assert result is not None
     assert result.pages == 5
@@ -149,7 +149,7 @@ async def test_get_metadata_returns_none_for_missing_field(
     maindb_driver: Driver, kbid: str, rid: str
 ) -> None:
     async with maindb_driver.ro_transaction() as txn:
-        result = await conversations_v2.get_metadata(txn, kbid=kbid, rid=rid, field_id="no-such-field")
+        result = await conversations.get_metadata(txn, kbid=kbid, rid=rid, field_id="no-such-field")
     assert result is None
 
 
@@ -163,11 +163,11 @@ async def test_set_and_get_page(maindb_driver: Driver, kbid: str, rid: str, fiel
     conv = make_conversation("Hello", "World")
 
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.set_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1, value=conv)
+        await conversations.set_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1, value=conv)
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        result = await conversations_v2.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1)
+        result = await conversations.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1)
 
     assert result is not None
     assert len(result.messages) == 2
@@ -178,19 +178,19 @@ async def test_set_and_get_page(maindb_driver: Driver, kbid: str, rid: str, fiel
 @pytest.mark.asyncio
 async def test_set_page_overwrites(maindb_driver: Driver, kbid: str, rid: str, field_id: str) -> None:
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.set_page(
+        await conversations.set_page(
             txn, kbid=kbid, rid=rid, field_id=field_id, page=1, value=make_conversation("v1")
         )
         await txn.commit()
 
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.set_page(
+        await conversations.set_page(
             txn, kbid=kbid, rid=rid, field_id=field_id, page=1, value=make_conversation("v2")
         )
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        result = await conversations_v2.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1)
+        result = await conversations.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1)
 
     assert result is not None
     assert result.messages[0].content.text == "v2"
@@ -201,17 +201,17 @@ async def test_multiple_pages_are_independent(
     maindb_driver: Driver, kbid: str, rid: str, field_id: str
 ) -> None:
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.set_page(
+        await conversations.set_page(
             txn, kbid=kbid, rid=rid, field_id=field_id, page=1, value=make_conversation("page-one")
         )
-        await conversations_v2.set_page(
+        await conversations.set_page(
             txn, kbid=kbid, rid=rid, field_id=field_id, page=2, value=make_conversation("page-two")
         )
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        p1 = await conversations_v2.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1)
-        p2 = await conversations_v2.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=2)
+        p1 = await conversations.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1)
+        p2 = await conversations.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=2)
 
     assert p1 is not None and p1.messages[0].content.text == "page-one"
     assert p2 is not None and p2.messages[0].content.text == "page-two"
@@ -222,7 +222,7 @@ async def test_get_page_returns_none_for_missing_page(
     maindb_driver: Driver, kbid: str, rid: str, field_id: str
 ) -> None:
     async with maindb_driver.ro_transaction() as txn:
-        result = await conversations_v2.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=99)
+        result = await conversations.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=99)
     assert result is None
 
 
@@ -232,7 +232,7 @@ async def test_set_page_rejects_page_zero(
 ) -> None:
     async with maindb_driver.rw_transaction() as txn:
         with pytest.raises(ValueError, match="pages start at index 1"):
-            await conversations_v2.set_page(
+            await conversations.set_page(
                 txn, kbid=kbid, rid=rid, field_id=field_id, page=0, value=make_conversation("x")
             )
 
@@ -243,7 +243,7 @@ async def test_get_page_rejects_page_zero(
 ) -> None:
     async with maindb_driver.ro_transaction() as txn:
         with pytest.raises(ValueError, match="pages start at index 1"):
-            await conversations_v2.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=0)
+            await conversations.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=0)
 
 
 # ---------------------------------------------------------------------------
@@ -258,13 +258,13 @@ async def test_set_and_get_splits_metadata(
     splits = make_splits_metadata("split-a", "split-b")
 
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.set_splits_metadata(
+        await conversations.set_splits_metadata(
             txn, kbid=kbid, rid=rid, field_id=field_id, splits_metadata=splits
         )
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        result = await conversations_v2.get_splits_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
+        result = await conversations.get_splits_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
 
     assert result is not None
     assert "split-a" in result.metadata
@@ -276,7 +276,7 @@ async def test_get_splits_metadata_returns_none_when_not_set(
     maindb_driver: Driver, kbid: str, rid: str, field_id: str
 ) -> None:
     async with maindb_driver.ro_transaction() as txn:
-        result = await conversations_v2.get_splits_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
+        result = await conversations.get_splits_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
     assert result is None
 
 
@@ -286,21 +286,21 @@ async def test_splits_metadata_does_not_clash_with_pages(
 ) -> None:
     """The sentinel page=0 must not conflict with real page data."""
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.set_splits_metadata(
+        await conversations.set_splits_metadata(
             txn,
             kbid=kbid,
             rid=rid,
             field_id=field_id,
             splits_metadata=make_splits_metadata("s1"),
         )
-        await conversations_v2.set_page(
+        await conversations.set_page(
             txn, kbid=kbid, rid=rid, field_id=field_id, page=1, value=make_conversation("msg")
         )
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        splits = await conversations_v2.get_splits_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
-        page = await conversations_v2.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1)
+        splits = await conversations.get_splits_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
+        page = await conversations.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1)
 
     assert splits is not None and "s1" in splits.metadata
     assert page is not None and page.messages[0].content.text == "msg"
@@ -316,13 +316,13 @@ async def test_delete_field_removes_all_pages_and_metadata(
     maindb_driver: Driver, kbid: str, rid: str, field_id: str
 ) -> None:
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.set_page(
+        await conversations.set_page(
             txn, kbid=kbid, rid=rid, field_id=field_id, page=1, value=make_conversation("p1")
         )
-        await conversations_v2.set_page(
+        await conversations.set_page(
             txn, kbid=kbid, rid=rid, field_id=field_id, page=2, value=make_conversation("p2")
         )
-        await conversations_v2.set_splits_metadata(
+        await conversations.set_splits_metadata(
             txn,
             kbid=kbid,
             rid=rid,
@@ -332,26 +332,21 @@ async def test_delete_field_removes_all_pages_and_metadata(
         await txn.commit()
 
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.delete_field(txn, kbid=kbid, rid=rid, field_id=field_id)
+        await conversations.delete_field(txn, kbid=kbid, rid=rid, field_id=field_id)
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
+        assert await conversations.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1) is None
+        assert await conversations.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=2) is None
         assert (
-            await conversations_v2.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=1) is None
-        )
-        assert (
-            await conversations_v2.get_page(txn, kbid=kbid, rid=rid, field_id=field_id, page=2) is None
-        )
-        assert (
-            await conversations_v2.get_splits_metadata(txn, kbid=kbid, rid=rid, field_id=field_id)
-            is None
+            await conversations.get_splits_metadata(txn, kbid=kbid, rid=rid, field_id=field_id) is None
         )
 
 
 @pytest.mark.asyncio
 async def test_delete_field_noop_when_no_rows_exist(maindb_driver: Driver, kbid: str, rid: str) -> None:
     async with maindb_driver.rw_transaction() as txn:
-        await conversations_v2.delete_field(txn, kbid=kbid, rid=rid, field_id="ghost")
+        await conversations.delete_field(txn, kbid=kbid, rid=rid, field_id="ghost")
         await txn.commit()  # must not raise
 
 
@@ -370,7 +365,7 @@ async def test_set_metadata_missing_parent_logs_warning(
 
     with caplog.at_level(logging.WARNING, logger="nucliadb.common.datamanagers.utils"):
         async with maindb_driver.rw_transaction() as txn:
-            await conversations_v2.set_metadata(
+            await conversations.set_metadata(
                 txn,
                 kbid=kbid,
                 rid=rid,
@@ -392,7 +387,7 @@ async def test_set_page_missing_parent_logs_warning(
 
     with caplog.at_level(logging.WARNING, logger="nucliadb.common.datamanagers.utils"):
         async with maindb_driver.rw_transaction() as txn:
-            await conversations_v2.set_page(
+            await conversations.set_page(
                 txn,
                 kbid=kbid,
                 rid=rid,
