+++ nucliadb/tests/nucliadb/integration/common/datamanagers/test_fields_v2.py
@@ -18,7 +18,7 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 """
-Integration tests for nucliadb.common.datamanagers.fields_v2.
+Integration tests for nucliadb.common.datamanagers.fields.
 
 Covers: set, set_status, get_raw, get_status, get_statuses, get_all_field_ids,
         has_field, delete.
@@ -28,7 +28,7 @@ import logging
 
 import pytest
 
-from nucliadb.common.datamanagers import fields_v2, kb_v2, resources_v2
+from nucliadb.common.datamanagers import fields, kb, resources
 from nucliadb.common.maindb.driver import Driver
 from nucliadb.ingest.orm.knowledgebox import KnowledgeBox
 from nucliadb.ingest.orm.resource import Resource
@@ -60,7 +60,7 @@ def make_status(
 async def kbid(maindb_driver: Driver) -> str:
     kbid = KnowledgeBox.new_unique_kbid()
     async with maindb_driver.rw_transaction() as txn:
-        await kb_v2.set_kbid_for_slug(txn, kbid=kbid, slug=f"slug-{kbid}")
+        await kb.set_kbid_for_slug(txn, kbid=kbid, slug=f"slug-{kbid}")
         await txn.commit()
     return kbid
 
@@ -69,7 +69,7 @@ async def kbid(maindb_driver: Driver) -> str:
 async def rid(maindb_driver: Driver, kbid: str) -> str:
     rid = Resource.new_unique_rid()
     async with maindb_driver.rw_transaction() as txn:
-        await resources_v2.set_slug(txn, kbid=kbid, rid=rid, slug=f"slug-{rid}")
+        await resources.set_slug(txn, kbid=kbid, rid=rid, slug=f"slug-{rid}")
         await txn.commit()
     return rid
 
@@ -84,11 +84,11 @@ async def test_set_and_get_raw(maindb_driver: Driver, kbid: str, rid: str) -> No
     payload = b"raw-field-bytes"
 
     async with maindb_driver.rw_transaction() as txn:
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=payload)
+        await fields.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=payload)
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        result = await fields_v2.get_raw(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body")
+        result = await fields.get_raw(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body")
 
     assert result == payload
 
@@ -96,15 +96,15 @@ async def test_set_and_get_raw(maindb_driver: Driver, kbid: str, rid: str) -> No
 @pytest.mark.asyncio
 async def test_set_overwrites_existing_value(maindb_driver: Driver, kbid: str, rid: str) -> None:
     async with maindb_driver.rw_transaction() as txn:
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"v1")
+        await fields.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"v1")
         await txn.commit()
 
     async with maindb_driver.rw_transaction() as txn:
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"v2")
+        await fields.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"v2")
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        result = await fields_v2.get_raw(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body")
+        result = await fields.get_raw(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body")
 
     assert result == b"v2"
 
@@ -114,9 +114,7 @@ async def test_get_raw_returns_none_for_missing_field(
     maindb_driver: Driver, kbid: str, rid: str
 ) -> None:
     async with maindb_driver.ro_transaction() as txn:
-        result = await fields_v2.get_raw(
-            txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="nonexistent"
-        )
+        result = await fields.get_raw(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="nonexistent")
     assert result is None
 
 
@@ -125,11 +123,11 @@ async def test_set_with_protobuf_message(maindb_driver: Driver, kbid: str, rid: 
     text_field = rpb2.FieldText(body="hello world", format=rpb2.FieldText.Format.PLAIN)
 
     async with maindb_driver.rw_transaction() as txn:
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=text_field)
+        await fields.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=text_field)
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        raw = await fields_v2.get_raw(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body")
+        raw = await fields.get_raw(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body")
 
     assert raw is not None
     recovered = rpb2.FieldText()
@@ -145,8 +143,8 @@ async def test_set_with_protobuf_message(maindb_driver: Driver, kbid: str, rid: 
 @pytest.mark.asyncio
 async def test_set_and_get_status(maindb_driver: Driver, kbid: str, rid: str) -> None:
     async with maindb_driver.rw_transaction() as txn:
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"data")
-        await fields_v2.set_status(
+        await fields.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"data")
+        await fields.set_status(
             txn,
             kbid=kbid,
             rid=rid,
@@ -157,7 +155,7 @@ async def test_set_and_get_status(maindb_driver: Driver, kbid: str, rid: str) ->
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        result = await fields_v2.get_status(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body")
+        result = await fields.get_status(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body")
 
     assert result is not None
     assert result.status == wpb2.FieldStatus.Status.PROCESSED
@@ -168,7 +166,7 @@ async def test_get_status_returns_none_for_missing_row(
     maindb_driver: Driver, kbid: str, rid: str
 ) -> None:
     async with maindb_driver.ro_transaction() as txn:
-        result = await fields_v2.get_status(
+        result = await fields.get_status(
             txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="no-such-field"
         )
     assert result is None
@@ -178,8 +176,8 @@ async def test_get_status_returns_none_for_missing_row(
 async def test_get_statuses_returns_in_order(maindb_driver: Driver, kbid: str, rid: str) -> None:
     async with maindb_driver.rw_transaction() as txn:
         for fid in ("f1", "f2", "f3"):
-            await fields_v2.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id=fid, value=b"x")
-        await fields_v2.set_status(
+            await fields.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id=fid, value=b"x")
+        await fields.set_status(
             txn,
             kbid=kbid,
             rid=rid,
@@ -187,7 +185,7 @@ async def test_get_statuses_returns_in_order(maindb_driver: Driver, kbid: str, r
             field_id="f1",
             status=make_status(wpb2.FieldStatus.Status.PROCESSED),
         )
-        await fields_v2.set_status(
+        await fields.set_status(
             txn,
             kbid=kbid,
             rid=rid,
@@ -204,7 +202,7 @@ async def test_get_statuses_returns_in_order(maindb_driver: Driver, kbid: str, r
     ]
 
     async with maindb_driver.ro_transaction() as txn:
-        statuses = await fields_v2.get_statuses(txn, kbid=kbid, rid=rid, fields=field_ids)
+        statuses = await fields.get_statuses(txn, kbid=kbid, rid=rid, fields=field_ids)
 
     assert len(statuses) == 3
     assert statuses[0].status == wpb2.FieldStatus.Status.PROCESSED  # f1
@@ -215,7 +213,7 @@ async def test_get_statuses_returns_in_order(maindb_driver: Driver, kbid: str, r
 @pytest.mark.asyncio
 async def test_get_statuses_empty_input(maindb_driver: Driver, kbid: str, rid: str) -> None:
     async with maindb_driver.ro_transaction() as txn:
-        result = await fields_v2.get_statuses(txn, kbid=kbid, rid=rid, fields=[])
+        result = await fields.get_statuses(txn, kbid=kbid, rid=rid, fields=[])
     assert result == []
 
 
@@ -227,19 +225,19 @@ async def test_get_statuses_empty_input(maindb_driver: Driver, kbid: str, rid: s
 @pytest.mark.asyncio
 async def test_has_field_true_after_set(maindb_driver: Driver, kbid: str, rid: str) -> None:
     async with maindb_driver.rw_transaction() as txn:
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"x")
+        await fields.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"x")
         await txn.commit()
 
     fid = rpb2.FieldID(field_type=rpb2.FieldType.TEXT, field="body")
     async with maindb_driver.ro_transaction() as txn:
-        assert await fields_v2.has_field(txn, kbid=kbid, rid=rid, field_id=fid) is True
+        assert await fields.has_field(txn, kbid=kbid, rid=rid, field_id=fid) is True
 
 
 @pytest.mark.asyncio
 async def test_has_field_false_for_missing(maindb_driver: Driver, kbid: str, rid: str) -> None:
     fid = rpb2.FieldID(field_type=rpb2.FieldType.TEXT, field="ghost")
     async with maindb_driver.ro_transaction() as txn:
-        assert await fields_v2.has_field(txn, kbid=kbid, rid=rid, field_id=fid) is False
+        assert await fields.has_field(txn, kbid=kbid, rid=rid, field_id=fid) is False
 
 
 # ---------------------------------------------------------------------------
@@ -250,15 +248,15 @@ async def test_has_field_false_for_missing(maindb_driver: Driver, kbid: str, rid
 @pytest.mark.asyncio
 async def test_get_all_field_ids(maindb_driver: Driver, kbid: str, rid: str) -> None:
     async with maindb_driver.rw_transaction() as txn:
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"x")
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type=FILE, field_id="doc", value=b"y")
+        await fields.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"x")
+        await fields.set(txn, kbid=kbid, rid=rid, field_type=FILE, field_id="doc", value=b"y")
         # These are the special title/summary generic fields that should be excluded
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type="a", field_id="title", value=b"t")
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type="a", field_id="summary", value=b"s")
+        await fields.set(txn, kbid=kbid, rid=rid, field_type="a", field_id="title", value=b"t")
+        await fields.set(txn, kbid=kbid, rid=rid, field_type="a", field_id="summary", value=b"s")
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        result = await fields_v2.get_all_field_ids(txn, kbid=kbid, rid=rid)
+        result = await fields.get_all_field_ids(txn, kbid=kbid, rid=rid)
 
     field_pairs = {(f.field_type, f.field) for f in result.fields}
     assert (rpb2.FieldType.TEXT, "body") in field_pairs
@@ -271,7 +269,7 @@ async def test_get_all_field_ids(maindb_driver: Driver, kbid: str, rid: str) -> 
 @pytest.mark.asyncio
 async def test_get_all_field_ids_empty(maindb_driver: Driver, kbid: str, rid: str) -> None:
     async with maindb_driver.ro_transaction() as txn:
-        result = await fields_v2.get_all_field_ids(txn, kbid=kbid, rid=rid)
+        result = await fields.get_all_field_ids(txn, kbid=kbid, rid=rid)
     assert list(result.fields) == []
 
 
@@ -283,23 +281,23 @@ async def test_get_all_field_ids_empty(maindb_driver: Driver, kbid: str, rid: st
 @pytest.mark.asyncio
 async def test_delete_removes_field(maindb_driver: Driver, kbid: str, rid: str) -> None:
     async with maindb_driver.rw_transaction() as txn:
-        await fields_v2.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"x")
+        await fields.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"x")
         await txn.commit()
 
     async with maindb_driver.rw_transaction() as txn:
-        await fields_v2.delete(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body")
+        await fields.delete(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body")
         await txn.commit()
 
     async with maindb_driver.ro_transaction() as txn:
-        assert await fields_v2.get_raw(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body") is None
+        assert await fields.get_raw(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body") is None
         fid = rpb2.FieldID(field_type=rpb2.FieldType.TEXT, field="body")
-        assert await fields_v2.has_field(txn, kbid=kbid, rid=rid, field_id=fid) is False
+        assert await fields.has_field(txn, kbid=kbid, rid=rid, field_id=fid) is False
 
 
 @pytest.mark.asyncio
 async def test_delete_nonexistent_field_is_noop(maindb_driver: Driver, kbid: str, rid: str) -> None:
     async with maindb_driver.rw_transaction() as txn:
-        await fields_v2.delete(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="ghost")
+        await fields.delete(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="ghost")
         await txn.commit()  # must not raise
 
 
@@ -319,7 +317,7 @@ async def test_set_missing_parent_logs_warning_and_does_not_raise(
 
     with caplog.at_level(logging.WARNING, logger="nucliadb.common.datamanagers.utils"):
         async with maindb_driver.rw_transaction() as txn:
-            await fields_v2.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"data")
+            await fields.set(txn, kbid=kbid, rid=rid, field_type=TEXT, field_id="body", value=b"data")
             await txn.commit()
 
     assert any("Foreign key violation" in r.message for r in caplog.records)
