+++ nucliadb/tests/search/integration/api/v1/test_suggest.py
@@ -23,11 +23,10 @@ from httpx import AsyncClient
 from nidx_protos.nodereader_pb2 import GetShardRequest, SuggestFeatures, SuggestRequest
 from nidx_protos.noderesources_pb2 import ShardId
 
-from nucliadb.common.datamanagers.cluster import KB_SHARDS
+from nucliadb.common.datamanagers.cluster import get_kb_shards
 from nucliadb.common.maindb.utils import get_driver
 from nucliadb.common.nidx import get_nidx_api_client, get_nidx_searcher_client
 from nucliadb.search.api.v1.router import KB_PREFIX
-from nucliadb_protos.writer_pb2 import Shards as PBShards
 
 
 @pytest.mark.deploy_modes("cluster")
@@ -45,24 +44,19 @@ async def test_suggest_resource_all(nucliadb_search: AsyncClient, test_search_re
 
     driver = get_driver()
     async with driver.ro_transaction() as txn:
-        key = KB_SHARDS.format(kbid=kbid)
-        async for key in txn.keys(key):
-            value = await txn.get(key)
-            assert value is not None
-            shards = PBShards()
-            shards.ParseFromString(value)
+        kb_shards = await get_kb_shards(txn, kbid=kbid)
+        assert kb_shards is not None
+        shard_id = kb_shards.shards[0].nidx_shard_id
+        shard = await get_nidx_api_client().GetShard(GetShardRequest(shard_id=ShardId(id=shard_id)))
+        assert shard.shard_id == shard_id
+        assert shard.fields == 3
+        assert shard.paragraphs == 2
+        assert shard.sentences == 3
 
-            shard_id = shards.shards[0].nidx_shard_id
-            shard = await get_nidx_api_client().GetShard(GetShardRequest(shard_id=ShardId(id=shard_id)))
-            assert shard.shard_id == shard_id
-            assert shard.fields == 3
-            assert shard.paragraphs == 2
-            assert shard.sentences == 3
-
-            prequest = SuggestRequest(
-                features=[SuggestFeatures.ENTITIES, SuggestFeatures.PARAGRAPHS],
-            )
-            prequest.shard_ids[:] = [shard_id]
-            prequest.body = "Ramon"
-            suggest = await get_nidx_searcher_client().Suggest(prequest)
-            assert suggest.total == 1, f"Request:\n{prequest}\nResponse:\n{suggest}"
+        prequest = SuggestRequest(
+            features=[SuggestFeatures.ENTITIES, SuggestFeatures.PARAGRAPHS],
+        )
+        prequest.shard_ids[:] = [shard_id]
+        prequest.body = "Ramon"
+        suggest = await get_nidx_searcher_client().Suggest(prequest)
+        assert suggest.total == 1, f"Request:\n{prequest}\nResponse:\n{suggest}"
