+++ nucliadb_utils/src/nucliadb_utils/const.py
@@ -38,15 +38,11 @@ class Features:
     SEMANTIC_GRAPH = "nucliadb_semantic_graph"
     AUDIT_RETRIEVE_AND_AUGMENT = "nucliadb_audit_retrieve_and_augment"
     # Flipt features
-    DATAMANAGERS_V2_READ = "nucliadb_datamanagers_v2_read"
-    DATAMANAGERS_V2_WRITE = "nucliadb_datamanagers_v2_write"
     BETTER_MULTI_SHARD_SEARCH = "nucliadb_better_multi_shard_search"
 
 
 _FliptFeatures = set(
     [
-        Features.DATAMANAGERS_V2_READ,
-        Features.DATAMANAGERS_V2_WRITE,
         Features.BETTER_MULTI_SHARD_SEARCH,
     ]
 )
