+++ nucliadb_utils/src/nucliadb_utils/featureflagging.py
@@ -60,21 +60,9 @@ DEFAULT_FLAG_DATA: dict[str, Any] = {
         "rollout": 0,
         "variants": {"environment": ["local"]},
     },
-    const.Features.DATAMANAGERS_V2_WRITE: {
+    const.Features.BETTER_MULTI_SHARD_SEARCH: {
         "rollout": 0,
-        "variants": {
-            "environment": ["local"],
-        },
-    },
-    const.Features.DATAMANAGERS_V2_READ: {
-        "rollout": 0,
-        "variants": {
-            "environment": ["local"],
-        },
-        const.Features.BETTER_MULTI_SHARD_SEARCH: {
-            "rollout": 0,
-            "variants": {"environment": ["local"]},
-        },
+        "variants": {"environment": ["local"]},
     },
 }
